hashindex: moved the tests from hashindex_stress to hashindex_pytest
